server: Provider mirror response to specific version requests
This is now a minimal valid implementation of Terraform's provider mirror
protocol, although it only really works for OCI Distribution registries
that don't require authentication because Terraform doesn't expect to need
to send credentials when downloading the actual package.

To make this work with registries that require auth will probably require
proxying the downloads through our own server, and packing enough
information into the URL to allow us to authenticate to the backend
without direcly disclosing the client's credentials in cleartext in the
URL. Maybe we'll do something more with that in later commits.