querysecret: Utility for sending secrets in a URL query string
The Terraform registry protocols expect actual artifacts to be delivered
from endpoints that don't require authentication, and so if we want to
authenticate those requests we need to place the necessary information
into the URL returned in the registry/mirror indices.

This package encapsulates the process of encrypting and authenticating
arbitrary small messages that expire after a few minutes, so that the
impact of them being exposed in e.g. access logs is minimized.

In real systems the server should also use TLS so that these messages will
not be visible to intermediaries at all. However, Terraform itself does
not understand these URLs as being secret and so it might disclose them
in its own UI while downloading artifacts, and so the encryption and
expiration together limit the window in which it would be useful to replay
those URLs to the server.