server: Optionally proxy package downloads through our server
This deals with the problem that the OCI Distribution registry will tend
to require the same credentials to retrieve a blob as they do to retrieve
the manifest that identified the blob, whereas Terraform's mirror protocol
assumes that the packages themselves do not need authentication.

This packs the blob digest and the necessary credentials into the query
string of the download URL as a time-limited encrypted message.